requirements.txt
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.5
charset-normalizer==3.4.4
colorama==0.4.6
Django==4.2.23
django-bootstrap-v5==1.0.11
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-leaflet==0.32.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
drf-extra-fields==3.7.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
filetype==1.2.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
leaflet==0.0.3
openpyxl==3.1.5
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pymongo==4.13.2
PyMySQL==1.1.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
reportlab==4.4.4
rpds-py==0.27.1
soupsieve==2.7
sqlparse==0.5.3
tailwindcss==0.0.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
xlwt==1.3.0